feat: support decimal quantities in POS operations
- Updated quantity handling in sales and returns to accept decimal values instead of only integers.
- Modified database schema to change relevant columns from INTEGER to REAL for products, sales lines, and inventory events.
- Enhanced barcode scanning functionality to recognize and process decimal quantities and prices in various formats.
- Improved user interface to reflect changes in quantity input and display appropriate messages for decimal values.
- Updated tests to cover new functionality for decimal quantities and modified product codes.

+++ packages/electron/src/offline/localDB.ts
@@ -322,7 +322,7 @@ function initSchema(db: Database.Database): void {
       subcategory TEXT NOT NULL DEFAULT '',
       pack_size INTEGER NOT NULL DEFAULT 1,
       unit_label TEXT NOT NULL DEFAULT 'pcs',
-      stock_on_hand INTEGER NOT NULL DEFAULT 0,
+      stock_on_hand REAL NOT NULL DEFAULT 0,
       description TEXT,
       variants_json TEXT,
       last_vector_clock TEXT DEFAULT '{}'
@@ -394,7 +394,7 @@ function initSchema(db: Database.Database): void {
       variant_name TEXT,
       variant_attributes_json TEXT,
       subcategory TEXT NOT NULL,
-      quantity INTEGER NOT NULL,
+      quantity REAL NOT NULL,
       unit_price REAL NOT NULL,
       tier_label TEXT NOT NULL,
       discount_percent REAL NOT NULL DEFAULT 0,
@@ -438,7 +438,7 @@ function initSchema(db: Database.Database): void {
       variant_name TEXT,
       variant_attributes_json TEXT,
       subcategory TEXT NOT NULL,
-      quantity INTEGER NOT NULL,
+      quantity REAL NOT NULL,
       unit_price REAL NOT NULL,
       tier_label TEXT NOT NULL,
       discount_percent REAL NOT NULL DEFAULT 0,
@@ -479,7 +479,7 @@ function initSchema(db: Database.Database): void {
       sale_line_id TEXT NOT NULL,
       product_id TEXT NOT NULL,
       variant_id TEXT,
-      quantity INTEGER NOT NULL,
+      quantity REAL NOT NULL,
       refund_amount REAL NOT NULL
     );
